docs/requirements.txt
h5py
matplotlib
nanite
numpy
pyqt5
scipy
# Remove sphinx from requirements b/c it lead to issues
# on rtd and it's a dependency of the rtd theme anyway.
# sphinx>=1.6.4,<2
sphinxcontrib.bibtex>=2.0
sphinx_rtd_theme
